Wrap classes into package
