+++ src/main/java/jarvis/DukeException.java
@@ -1,3 +1,5 @@
+package jarvis;
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
