+++ src/main/java/jarvis/Event.java
@@ -1,28 +1,33 @@
+package jarvis;
+
+import jarvis.task.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task {
-    private LocalDateTime by;
-    public Deadline(String input, boolean isDone) throws DateTimeParseException {
+public class Event extends Task {
+
+    private LocalDateTime at;
+    public Event(String input, boolean isDone) throws DateTimeParseException {
         super(isDone);
-        String[] strArr = input.split("/by");
+        String[] strArr = input.split("/at");
         this.description = strArr[0].trim();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
         try {
-            this.by = LocalDateTime.parse(strArr[1].trim(), formatter);
+            this.at = LocalDateTime.parse(strArr[1].trim(), formatter);
         } catch (DateTimeParseException e) {
             System.out.println("Wrong date format");
             throw e;
         }
     }
 
-    public Deadline(String input, String by, boolean isDone) {
+    public Event(String input, String at, boolean isDone) {
         super(isDone);
         this.description = input;
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
         try {
-            this.by = LocalDateTime.parse(by);
+            this.at = LocalDateTime.parse(at);
         } catch (DateTimeParseException e) {
             System.out.println("Wrong date format");
             throw e;
@@ -31,20 +36,20 @@ public class Deadline extends Task {
 
     @Override
     public TaskType getTaskType() {
-        return TaskType.Deadline;
+        return TaskType.Event;
     }
 
     @Override
     public String toDataForm() {
         String done = this.isDone ? "1" : "0";
-        return "D|" + done + "|" + this.description + "|" + this.by + "\n";
+        return "D|" + done + "|" + this.description + "|" + this.at + "\n";
     }
 
     @Override
     public String toString() {
-        String date = this.by.format(DateTimeFormatter.ofPattern("dd MMM yyyy HH:mm"));
-        String head = "[D][" + this.getStatusIcon() + "] ";
-        String body = this.description + " (by: " + date + ")";
+        String date  = this.at.format(DateTimeFormatter.ofPattern("dd MMM yyyy HH:mm"));
+        String head = "[E][" + this.getStatusIcon() + "] ";
+        String body = this.description + " (at: " + date + ")";
         return head + body;
     }
 }
