+++ src/main/java/jarvis/Ui.java
@@ -1,3 +1,7 @@
+package jarvis;
+
+import jarvis.task.TaskList;
+
 import java.util.Scanner;
 
 public class Ui {
