+++ src/main/java/jarvis/task/Todo.java
@@ -1,3 +1,5 @@
+package jarvis.task;
+
 public class Todo extends Task {
     public Todo(String input, boolean isDone) {
         super(isDone);
